refactor: department class changed package location, integration test class extends
