+++ src/main/kotlin/com/tistory/shanepark/dutypark/common/advice/RestExceptionControllerAdvice.kt
@@ -3,7 +3,6 @@ package com.tistory.shanepark.dutypark.common.advice
 import com.tistory.shanepark.dutypark.common.exceptions.DutyparkAuthException
 import org.springframework.core.Ordered
 import org.springframework.core.annotation.Order
-import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 
@@ -11,7 +10,6 @@ import org.springframework.web.bind.annotation.*
 @Order(Ordered.HIGHEST_PRECEDENCE)
 class RestExceptionControllerAdvice {
 
-    @ResponseStatus(HttpStatus.OK)
     @ResponseBody
     @ExceptionHandler
     fun notAuthorizedHandler(e: DutyparkAuthException): ResponseEntity<Any> {
