--- src/main/kotlin/com/tistory/shanepark/dutypark/member/repository/DepartmentRepository.kt
@@ -1,7 +1,0 @@
-package com.tistory.shanepark.dutypark.member.repository
-
-import com.tistory.shanepark.dutypark.member.domain.entity.Department
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface DepartmentRepository : JpaRepository<Department, Long> {
-}
