+++ src/test/kotlin/com/tistory/shanepark/dutypark/DutyparkIntegrationTest.kt
@@ -1,0 +1,8 @@
+package com.tistory.shanepark.dutypark
+
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.transaction.annotation.Transactional
+
+@SpringBootTest
+@Transactional
+class DutyparkIntegrationTest
