+++ src/test/kotlin/com/tistory/shanepark/dutypark/common/WebhookTest.kt
@@ -1,18 +1,15 @@
 package com.tistory.shanepark.dutypark.common
 
+import com.tistory.shanepark.dutypark.DutyparkIntegrationTest
 import net.gpedro.integrations.slack.SlackApi
 import net.gpedro.integrations.slack.SlackMessage
 import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Value
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.test.context.ActiveProfiles
 
-@SpringBootTest
-@ActiveProfiles("dev")
 @Disabled("token is required")
-class WebhookTest {
+class WebhookTest : DutyparkIntegrationTest() {
 
     @Value("\${dutypark.slack.token}")
     lateinit var token: String
