+++ src/test/kotlin/com/tistory/shanepark/dutypark/security/domain/entity/RefreshTokenTest.kt
@@ -1,6 +1,6 @@
 package com.tistory.shanepark.dutypark.security.domain.entity
 
-import com.tistory.shanepark.dutypark.member.domain.entity.Department
+import com.tistory.shanepark.dutypark.department.domain.entity.Department
 import com.tistory.shanepark.dutypark.member.domain.entity.Member
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
